Refactor tests/: Move `get_tmp_path, gen_path` to `common.rs`
